requirements.txt
autopep8==1.3.2
boto==2.48.0
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
croniter==0.3.20
diff-match-patch==20121119
django==2.1.4
django-cachalot==1.5.0
django-cors-headers==2.1.0
django-crontab==0.7.1
django-debug-toolbar==1.8
django-extensions==1.8.1
django-filter==1.0.4
django-import-export==0.5.1
django-modeladmin-reorder==0.3
django-redis==4.8.0
django-rest-swagger==2.1.2
django-silk==1.0.0
django-simple-history==1.9.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
fpdf==1.7.2
gunicorn==19.7.1
idna==2.5
img2pdf==0.2.4
itypes==1.1.0
jdcal==1.3
Jinja2==2.9.6
jsondiff==1.1.1
MarkupSafe==1.0
mongoengine==0.15.3
MonthDelta==0.9.1
mysqlclient==1.3.10
numpy==1.14.2
odfpy==1.3.5
openapi-codec==1.3.2
openpyxl==2.4.8
pandas==0.21.1
pdfkit==0.6.1
Pillow==5.0.0
pyBarcode==0.7
pycodestyle==2.3.1
Pygments==2.2.0
PyJWT==1.5.2
pymongo==3.7.1
python-barcode==0.8.3
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
redis==2.10.5
reportlab==3.4.0
requests==2.18.3
simplejson==3.13.2
six==1.10.0
sqlparse==0.2.3
tablib==0.11.5
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
uuid==1.30
validate-email==1.3
xlrd==1.0.0
xlwt==1.2.0
